server/requirements.txt
cassandra-driver==3.25.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
geomet==0.2.1.post1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
MarkupSafe==2.0.1
py4j==0.10.9
pyspark==3.1.2
requests==2.25.1
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
